+++ src/Commands/User/UserCreateCommand.php
@@ -4,6 +4,7 @@ use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\Question;
 use Wireshell\Helpers\PwUserTools;
 
 /**
@@ -13,6 +14,7 @@ use Wireshell\Helpers\PwUserTools;
  *
  * @package Wireshell
  * @author Marcus Herrmann
+ * @author Tabea David
  */
 class UserCreateCommand extends PwUserTools
 {
@@ -41,21 +43,31 @@ class UserCreateCommand extends PwUserTools
         parent::bootstrapProcessWire($output);
 
         $name = $input->getArgument('name');
-        $roles = explode(",", $input->getOption('roles'));
+        $email = $input->getOption('email');
         $pass = $input->getOption('password');
+        $roles = explode(",", $input->getOption('roles'));
+
+        $helper = $this->getHelper('question');
+        if (!$email) {
+            $question = new Question('Please enter a email address : ', 'email');
+            $email = $helper->ask($input, $output, $question);
+        }
+        if (!$pass) {
+            $question = new Question('Please enter a password : ', 'password');
+            $question->setHidden(true);
+            $question->setHiddenFallback(false);
+            $pass = $helper->ask($input, $output, $question);
+        }
 
         if (!wire("pages")->get("name={$name}") instanceof \NullPage) {
-
             $output->writeln("<error>User '{$name}' already exists!</error>");
             exit(1);
         }
 
-        $user = $this->createUser($input, $name, $this->userContainer, $pass);
+        $user = $this->createUser($email, $name, $this->userContainer, $pass);
         $user->save();
 
-        if ($input->getOption('roles')) {
-            $this->attachRolesToUser($name, $roles, $output);
-        }
+        if ($input->getOption('roles')) $this->attachRolesToUser($name, $roles, $output);
 
         if ($pass) {
           $output->writeln("<info>User '{$name}' created successfully!</info>");
@@ -63,6 +75,5 @@ class UserCreateCommand extends PwUserTools
           $output->writeln("<info>User '{$name}' created successfully! Please do not forget to set a password.</info>");
         }
     }
-
 
 }